chore: remove deprecated jx console cmd
fixes #6928
