+++ pkg/cmd/cmd.go
@@ -169,7 +169,6 @@ func NewJXCommand(f clients.Factory, in terminal.FileReader, out terminal.FileWr
 		{
 			Message: "Working with Applications:",
 			Commands: []*cobra.Command{
-				NewCmdConsole(commonOpts),
 				NewCmdLogs(commonOpts),
 				NewCmdOpen(commonOpts),
 				rsh.NewCmdRsh(commonOpts),
